requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-RESTful==0.3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
